Refactor JsonConfigurationProvider into focused partial files
Split the 567-line JsonConfigurationProvider.cs into four focused
partial-class files, each under 300 lines, per issue #9983:

- JsonConfigurationProvider.Loading.cs: constructor, fields, logging
  helpers, LoadAsync, TryGet.
- JsonConfigurationProvider.Sections.cs: GetSection and
  ThrowSectionMustBeAnObject.
- JsonConfigurationProvider.CommandLineOptions.cs:
  EnumerateCommandLineOptions and ThrowEntryMustBeScalarOrArray.
- JsonConfigurationProvider.Normalization.cs:
  NormalizeCommandLineOptionScalars, StartsWithChar, OptionBuilder.

Behavior and internal API are unchanged.

Copilot-Session: 79da38c3-8f37-4af7-8744-d14cd62402c3
